Product controller tests

+++ product-api/src/main/java/com/qima/product/product/application/command/ProductCommand.java
@@ -1,5 +1,7 @@
 package com.qima.product.product.application.command;
 
+import java.math.BigDecimal;
+
 public record ProductCommand(
         Long id,
         String name,
@@ -7,5 +9,5 @@ public record ProductCommand(
         String categoryPath,
         Long categoryId,
         Boolean available,
-        Double price) {
+        BigDecimal price) {
 }
